+++ tests/adaptors/test_terachem.py
@@ -1,7 +1,6 @@
 from pathlib import Path
 
 import pytest
-from qccodec.encoders.terachem import XYZ_FILENAME
 
 from qccompute.adapters import TeraChemAdapter
 from qccompute.adapters.utils import tmpdir
@@ -49,7 +48,7 @@ def test_propagate_wfn(prog_input_factory, results):
         adapter.propagate_wfn(results, prog_input_factory_inst)
 
     # TeraChem Output conventions
-    scr_postfix = XYZ_FILENAME.split(".")[0]
+    scr_postfix = "geometry".split(".")[0]
     scr_dir = f"scr.{scr_postfix}"
 
     # Add restricted wavefunction data to output
@@ -86,7 +85,7 @@ def test_collect_wfn(results):
         adapter.collect_wfn()
 
     # Check collection of c0
-    scr_dir_str = f"scr.{XYZ_FILENAME.split('.')[0]}"
+    scr_dir_str = "scr.geometry"
     with tmpdir():
         scr_dir = Path(scr_dir_str)
         scr_dir.mkdir()